Started working on the discrete cosine transform
I decided that I would put all the reds in a vector all the greens in a vector and all the blues in a vector. Then I would take the dct of each vector and concatenate all the vectors. I'm honestly not sure how to impliment this in a pca but I will look further into it.